requirements.txt
requests==2.31.0
python-dotenv==1.0.0
crewai>=0.51.0
composio-crewai==0.7.8 